Made case slider & algorithm selection as a general settings for all graphs
